Extracted TestAbstractFormatter into a separate file.
